+++ releasenotes.md
@@ -16,6 +16,8 @@
 - Fixed: regression in comments for `@deprecated` and `@pure`.
 - Detect recursive creation of generics #2366.
 - Compiler assertion when defining a function with return type untyped_list #2368.
+- Compiler assert when using generic parameters list without any parameters. #2369
+- Parsing difference between "0x00." and "0X00." literals #2371
 
 ### Stdlib changes
 - Add `==` to `Pair`, `Triple` and TzDateTime. Add print to `Pair` and `Triple`.